*-mainloop: add dummy render to unblock input
Newer versions of Android will stop delivering input to applications
that don't appear to be responsive because they aren't rendering
anything.

This updates na-mainloop and agdk-mainloop so that they at least
do a NOOP post of an untouched framebuffer each frame, so that they
continue to be delivered touch input.